pyproject.toml
[project]
name = "audio-classification"
version = "0.1.0"
description = "2 models 2 represetations = 4 experiments"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "librosa>=0.11.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "torch>=2.10.0",
    "typer>=0.24.1",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["infra*", "models*"]

[project.scripts]
audio-clf = "infra.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow: long-running tests"]
